move MultiHostReverseProxy logic away from the server command
